requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==26.1.0
babel==2.18.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.22.10
dbt-common==1.37.3
dbt-core==1.11.8
dbt-extractor==0.6.0
dbt-postgres==1.10.0
dbt-protos==1.0.443
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.2
Faker==40.13.0
idna==3.11
importlib_metadata==8.9.0
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.4.4
orderly-set==5.5.0
packaging==26.0
pandas==3.0.2
parsedatetime==2.6
pathspec==0.12.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
snowplow-tracker==1.1.0
sqlparse==0.5.5
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
